More restrictive type check for min/max values in CScriptableVars
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4664 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/client/DeprecatedGUI/Menu_Local.cpp
@@ -968,7 +968,7 @@ static void updateFeaturesList(CListview* l)
 						float fVal = atof(textBox->getText());
 						CLAMP_DIRECT(fVal, it->second.min.f, it->second.max.f );
 						* it->second.var.f = fVal;
-						slider->setValue(fVal * 10.0f);
+						slider->setValue(int(fVal * 10.0f));
 					}
 				}
 			}
